MNT: make VDS support conditional
Only expose the VDS related classes and method if VDS is available

 - move VDS classes to separate module
 - put Group.create_virtual behind a version guard
 - put top-level imports behind version guard
